EPMRPP-98254 || Code Review fix - 1
