+++ app/localization/translated/be.json
@@ -285,7 +285,6 @@
   "Common.friday": "Пятніца",
   "Common.generate": "Генераваць",
   "Common.invite": "Запрасіць",
-  "Common.inviteAndAssign": "Запрасіць і прызначыць",
   "Common.january": "Студзень",
   "Common.july": "Ліпень",
   "Common.june": "Чэрвень",
