+++ app/localization/translated/es.json
@@ -285,7 +285,6 @@
   "Common.friday": "Viernes",
   "Common.generate": "Generar",
   "Common.invite": "Invitar",
-  "Common.inviteAndAssign": "Invite & assign",
   "Common.january": "Enero",
   "Common.july": "Julio",
   "Common.june": "Junio",
