+++ app/localization/translated/ru.json
@@ -285,7 +285,6 @@
   "Common.friday": "Пятница",
   "Common.generate": "Сгенерировать",
   "Common.invite": "Пригласить",
-  "Common.inviteAndAssign": "Пригласить и назначить",
   "Common.january": "Январь",
   "Common.july": "Июль",
   "Common.june": "Июнь",
