+++ app/localization/translated/uk.json
@@ -285,7 +285,6 @@
   "Common.friday": "П’ятниця",
   "Common.generate": "Генерувати",
   "Common.invite": "Запросити",
-  "Common.inviteAndAssign": "Запросити і призначити",
   "Common.january": "Січень",
   "Common.july": "Липень",
   "Common.june": "Червень",
