+++ app/localization/translated/zh.json
@@ -285,7 +285,6 @@
   "Common.friday": "星期五",
   "Common.generate": "Generate",
   "Common.invite": "邀请",
-  "Common.inviteAndAssign": "Invite & assign",
   "Common.january": "一月",
   "Common.july": "七月",
   "Common.june": "六月",
