+++ app/src/common/constants/localization.js
@@ -87,10 +87,6 @@ export const COMMON_LOCALE_KEYS = defineMessages({
     id: 'Common.invite',
     defaultMessage: 'Invite',
   },
-  INVITE_AND_ASSIGN: {
-    id: 'Common.inviteAndAssign',
-    defaultMessage: 'Invite & assign',
-  },
   OK: {
     id: 'Common.ok',
     defaultMessage: 'Ok',
