+++ app/src/pages/common/membersPage/index.js
@@ -16,5 +16,4 @@
 
 export { MembersPage } from './membersPage';
 export { PermissionMapModal } from './modals/permissionMapModal';
-export { InviteUserModal } from './modals/inviteUserModal';
 export { ExternalUserInvitationModal } from './modals/externalUserInvitationModal';
